[feat] [관리자 권한관리] 권한별 매핑 Back API 구현

+++ hsb-bo/src/main/java/com/hbs/hsbbo/admin/domain/entity/AdminMenu.java
@@ -11,7 +11,7 @@ public class AdminMenu {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;  // 시퀀스값, AUTO_INCREMENT
+    private Long id;  // 시퀀스값, AUTO_INCREMENT
 
     @Column(nullable = false, length = 100)
     private String name;  // 메뉴명
